+++ tests/cli/fixtures/cli_integration_datasets.py
@@ -23,11 +23,11 @@ from renku.ui.cli import cli
 
 
 @pytest.fixture()
-def create_s3_dataset(runner):
+def create_cloud_storage_dataset(runner):
     """Fixture that creates a dataset given its name and uri."""
 
-    def _create_s3_dataset(name, uri):
+    def _create_cloud_storage_dataset(name, uri):
         res = runner.invoke(cli, ["dataset", "create", name, "--storage", uri])
         return res
 
-    yield _create_s3_dataset
+    yield _create_cloud_storage_dataset