+++ tests/cli/test_workflow_file.py
@@ -71,12 +71,15 @@ def test_dry_run_workflow_file(runner, workflow_file_project):
 
 def test_run_workflow_file_with_selected_steps(runner, workflow_file_project):
     """Test running a sub-set of steps of a workflow file."""
-    result = runner.invoke(cli, ["run", "--dry-run", workflow_file_project.workflow_file, "head", "line-count"])
+    result = runner.invoke(cli, ["run", workflow_file_project.workflow_file, "head", "tail"])
     assert 0 == result.exit_code, format_result_exception(result)
 
-    assert "Will execute step 'head': head $n $models $colors > $temporary-result" in result.output
-    assert "Will execute step 'tail': tail $parameters intermediate > results/output.csv" not in result.output
-    assert "Will execute step 'line-count': wc -l $models-and-colors > $output" in result.output
+    assert "Executing step 'workflow-file.head':" in result.output
+    assert "Executing step 'workflow-file.tail':" in result.output
+    assert "Executing step 'workflow-file.line-count':" not in result.output
+
+    # Third step's output isn't created
+    assert not (workflow_file_project.path / "results" / "output.csv.wc").exists()
 
 
 def test_run_workflow_file_with_no_commit(runner, workflow_file_project):
@@ -355,6 +358,41 @@ def test_workflow_file_plan_versioning(runner, workflow_file_project, with_injec
     assert line_count_1.derived_from is None
     assert line_count_2.derived_from is None
     assert line_count_3.derived_from is None
+
+
+def test_workflow_file_plan_versioning_with_selected_steps(runner, workflow_file_project, with_injection):
+    """Test plans are versioned correctly when executing subsets of steps."""
+    result = runner.invoke(cli, ["run", workflow_file_project.workflow_file, "head", "tail"])
+    assert 0 == result.exit_code, format_result_exception(result)
+    time.sleep(1)
+
+    with with_injection():
+        plan_gateway = PlanGateway()
+        root_plan_1 = plan_gateway.get_by_name("workflow-file")
+        head_1 = plan_gateway.get_by_name("workflow-file.head")
+        tail_1 = plan_gateway.get_by_name("workflow-file.tail")
+        line_count_1 = plan_gateway.get_by_name("workflow-file.line-count")
+
+    result = runner.invoke(cli, ["run", workflow_file_project.workflow_file])
+    assert 0 == result.exit_code, format_result_exception(result)
+
+    time.sleep(1)
+
+    with with_injection():
+        plan_gateway = PlanGateway()
+        root_plan_2 = plan_gateway.get_by_name("workflow-file")
+        head_2 = plan_gateway.get_by_name("workflow-file.head")
+        tail_2 = plan_gateway.get_by_name("workflow-file.tail")
+        line_count_2 = plan_gateway.get_by_name("workflow-file.line-count")
+
+    # Plan `line-count` wasn't executed in the first run
+    assert line_count_1 is None
+    assert line_count_2 is not None
+
+    # Everything else is the same
+    assert root_plan_2.id == root_plan_1.id
+    assert head_2.id == head_1.id
+    assert tail_2.id == tail_1.id
 
 
 def test_duplicate_workflow_file_plan_name(runner, workflow_file_project):